Refactor: Add version info, convert dropzone layout to grid, minor fixes
* Added footer with version info and github repo link, incentivizing to
contribute.
* Converted dropzone layout to css grid for the header, help, and footer
 sections to ensure their containers maintain the correct proportions
 when browser window is resized (responsive).
* Fixed an issue where on startup, if the H key is pressed to hide the
dropzone, it remains hidden when draggin images to it. This caused the
images to be opened by the browser as regular image files, each in a
separate window. Setup a conditional on H keypress to detect grid item
size and prevent hiding dropzone if size is `0`.
* Fixed an issue where on image delete from fullscreen preview, that
image is kept in memory. That same image is displayed on F keypress
once again even if previously deleted. Fixed by setting the
`view.currentImage` variable to empty on delete. This also prevents any
text from being copied on C keypress.
* When deleting, if all images are removed, the dropzone panel is
displayed again, including help.
* Swapped header's H3 to H1 for correct hierarchy naming standard.